(hel-668) Séparation des tests sur la recherche par numéro Finess
